requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
aiopg==1.4.0
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.2
Faker==22.2.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
multidict==6.0.4
psycopg2-binary==2.9.9
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
yarl==1.9.4
